LRCI-3823 Move available test suites into BaseGitHubEventHandler
